pyproject.toml
[project]
name = "berkshire-genai-signal-engine"
version = "0.1.0"
description = "Extract structured financial signals from Berkshire Hathaway shareholder letters using GenAI"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "yfinance>=0.2.40",
    "streamlit>=1.35",
    "plotly>=5.22",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
